Implement basic line scrolling
The main UI thread now prints a new line every second to a TextView.
When it reaches the bottom, the TextView scrolls down.

The current implementation iteratively concatenates a string which is
then printed out in its entirety before the scrolling fixes everything
up. Thus the computations necessary to print each new line increases
linearly in the line number. Since the printing is all being done in the
UI thread, the UI thread computation time increases linearly in the line
number, whereas the AsyncTask background thread's computation time is
constant in the line number.

This is obviously a bad implementation and it's not hypothetical, since
the UI thread slows down noticably in under 1000 lines. An obvious fix
is to use a sort of frame buffer which should contain a fixed upper
bound of text strings representing lines. This fix alone would make the
UI thread run in constant time in the line number.