Beginnings of basic terminal class
I have decided to hardcode the number of line of my terminal at 30 for
the moment. This only works on phones where this holds (which is true
for mine and for my emulator). This will not work once I start adding
buttons, but that will be dealt with when it arrives. Finding dimensions
of views is apparently pretty frustrating...

Some basic logic is still necessary for the Terminal. A public method of
the form Terminal.write(String str) should be added. Everything else
should be done under the hood.