Implement basic line Terminal
Now the Terminal class as a public method writeLine(). This is the only
way to add to the TextView. The current number of lines is hardcoded at
30 which works for my phone and emulator. Ideally the number of lines
should be discovered when the app starts up, but this is irritatingly
difficult to do right. There is a background line buffer which only
stores 30 lines so the updating of the UI is not constant in time.