Add BluetoothConn skeleton
Now I have a class BluetoothConn which I will use to connect to the
Mindset. At the moment it pretends to connect and then gives fake data
back when it is asked for. At this point, the rest of the work should go
into actually implementing the methods of the BluetoothConn class.