Add functionality to read from the device
The phone successfully is reading from the device, but it is doing no
parsing. Right now the connection being established is kind of a hack.
For some reason one needs to turn off Bluetooth on the phone before
running the app. Then the app turns on Bluetooth and everything works.
It does not seem to be releasing resources correctly or something.
Possibly the startup logic (where it checks that Bluetooth is enabled)
is busted as well.