Fix use of Bluetooth resources
Now the app connects to the Mindset upon start up and disconnects when
stopped or killed. After being stopped, the app reaquires the resources
correctly when starting back up. After being killed, it starts back up
from the beginning as it should. If any steps during startup or restart
up fail, an error description is printed and the app is halted entirely.