Remove BluetoothConn class
The functionality of the BluetoothConn class was merged back into the
main Activity class. The reason is because of all the setup and clean up
one must do with the Bluetooth connections when the Activity changes
state (onCreate(), onStart(), onStop(), etc.). Also the method
getBluetoothAdapter() needs to be run in the main Activity first
regardless because of a bug in Android.

Maybe in the future it will make sense to separate out this
functionality, but for the time being this seems to be the cleanest way
to get everything working well together.