Change MindsetParser class to act on InputStreams
The old MindsetParser class took a Bluetooth RFComm socket as input and
then immediately created and then acted upon the InputStream. Hence
having parser class know anything about Bluetooth makes no sense. It
should simply take a stream which provides a read method that returns a
byte (or possibly character). Maybe later it should be changed so that
the parser only requires an interface instead of a class.