Add correct parsing capabilities
The parser is now implemented. It does not follow all of the Neurosky
mindset specifications, but it implements the parsing I want
implemented. It only parses packet payloads of size 32 bytes. This
corresponds to a packet payload containing signal quality information,
meditation and concentration, as well as the beta/gamma/etc. waves. The
only other thing my Mindset will put out a period blink signal, but I
don't use that anyway.