Add complete app skeleton
The overall structure of the app is complete except for the
implementation of the logistic regression class (which is independent of
the structure anyway). There are, however, many runtime errors. There
seem to be problems resulting from the current need to close and reopen
the Bluetooth connections between the DataGathering and Predict
activities. Furthermore, there are large amounts of duplicated code.
Refractoring is surely necessary. Either the DataGathering and
Predicting screens should be combined (which would simplify many things
actually), or they should certainly inherent from a base class.